+++ app/[locale]/sitemap-categories.xml/route.ts
@@ -1,9 +1,11 @@
-import { Product } from 'shared/types/product/Product';
+import { NextRequest } from 'next/server';
+import { Category } from 'shared/types/product/Category';
 import { SiteMapField, generateSiteMap } from 'helpers/sitemap';
+import { i18nConfig } from 'project.config';
 import { sdk } from 'sdk';
 
-export async function GET() {
-  const locale = 'en';
+export async function GET(request: NextRequest, { params }: { params: { locale: string } }) {
+  const locale = params.locale ?? i18nConfig.defaultLocale;
 
   const siteUrl = process.env.SITE_URL;
 
@@ -13,26 +15,26 @@ export async function GET() {
 
   let nextCursor: string | undefined;
 
+  const extensions = sdk.composableCommerce;
+
   do {
-    const extensions = sdk.composableCommerce;
+    const response = await extensions.product.queryCategories({ cursor: nextCursor, limit: 12 });
 
-    const response = await extensions.product.query({ cursor: nextCursor, limit: 12 });
-
-    const items = [] as Product[];
+    const items = [] as Category[];
 
     if (!response.isError && response.data.items != null) {
       items.push(...response.data.items);
     }
 
     fields.push(
-      ...items?.map((product) => ({
-        loc: `${siteUrl}/${locale}${product._url}`,
+      ...items?.map((category) => ({
+        loc: `${siteUrl}/${locale}${category._url}`,
         lastmod: new Date().toISOString(),
         changefreq: 'daily' as const,
       })),
     );
 
-    nextCursor = !response.isError ? response.data.nextCursor : undefined;
+    nextCursor = (!response.isError && response.data.nextCursor) as string;
   } while (nextCursor);
 
   return new Response(generateSiteMap(fields), {