+++ components/commercetools-ui/organisms/product/product-list-algolia/components/facets/color.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRefinementList } from 'react-instantsearch-hooks-web';
-import stringToColor from 'string-to-color';
 import { FacetProps } from './types';
+import { textToColor } from 'helpers/textToColor/textToColor';
 
 const ColorFacet: React.FC<FacetProps> = ({ attribute }) => {
   const { items, refine } = useRefinementList({ attribute });
@@ -18,7 +18,7 @@ const ColorFacet: React.FC<FacetProps> = ({ attribute }) => {
             className={`h-40 w-40 rounded-full outline outline-1 outline-offset-1 ${
               isRefined ? 'outline-gray-500' : 'outline-transparent'
             }`}
-            style={{ backgroundColor: stringToColor(value) }}
+            style={{ backgroundColor: textToColor(value) }}
           />
           <span className="mt-4 block max-w-full truncate text-14" title={label}>
             {label}