+++ frontastic/renderer/components/tastic-wrapper/index.tsx
@@ -4,17 +4,20 @@ import { TasticWrapperProps } from './types';
 import { deviceVisibility } from '../../utils/device-visibility';
 import { highlight } from '../../utils/highlight';
 import { injectDataSources } from '../../utils/inject-datasources';
+import MissingTastic from '../missing-tastic';
 
 const TasticWrapper = ({ data, params, searchParams, dataSources, isHighlighted, categories }: TasticWrapperProps) => {
   const Tastic = tastics[data.tasticType];
-
-  if (!Tastic) return <></>;
 
   const resolvedTasticData = dataSources ? injectDataSources(data.configuration, dataSources) : data.configuration;
 
   return (
     <div id={data?.tasticId} className={`${highlight(isHighlighted)} ${deviceVisibility(data.configuration)}`}>
-      <Tastic data={resolvedTasticData} params={params} searchParams={searchParams} categories={categories} />
+      {Tastic ? (
+        <Tastic data={resolvedTasticData} params={params} searchParams={searchParams} categories={categories} />
+      ) : (
+        <MissingTastic data={data} />
+      )}
     </div>
   );
 };